Don't run tests in Docker
This container is only used for testing with Circle CI, which has its
own way of specifying installation steps. In addition, not only is this
currently broken due to the recent 'requirements.txt' -> 'setup.py'
changes on 'master', but this will break if/when the dependencies of
'stable' differ from 'master'.

Given that installing dependencies this way is both broken and
unnecessary, we can and should stop doing it.

Signed-off-by: Stephen Finucane <stephen@that.guru>